+++ mlf_core/create/templates/mlflow/mlflow_xgboost/{{ cookiecutter.project_slug }}/{{ cookiecutter.project_slug_no_hyphen }}/mlf_core/mlf_core.py
@@ -1,3 +1,5 @@
+import platform
+
 import click
 import tempfile
 import mlflow
@@ -20,6 +22,12 @@ def log_sys_intel_conda_env():
 
 
 def log_system_intelligence(reports_output_dir: str):
+    current_platform = platform.system()
+    if current_platform != 'Linux':
+        click.echo(click.style(f'Running on {current_platform} which is not supported by system-intelligence. Skipping hardware report.', fg='red'))
+        click.echo(click.style(f'Run MLflow with Docker to enforce a Linux environment.', fg='blue'))
+        return
+
     # Scoped import to prevent issues like RuntimeError: Numba cannot operate on non-primary CUDA context
     from system_intelligence.query import query_and_export